Centralize player valuation into PlayerValuationService
Consolidate fragmented ability↔market value conversion logic from
SeedReferenceData, PlayerDevelopmentService, and PlayerGeneratorService
into a single service. This fixes AI team market values becoming stale
after season 1 by recalculating values for ALL players during season-end
development. Uses pure ability average instead of overall_score so
fitness/morale no longer permanently affect valuation.

Remove dead ProcessSeasonDevelopment flow (action, command, event, route)
that was never referenced from any template.
